Torus approach (using simple geometry)
The calculation of P_vec (the vector of the previous segment) was moved from the loop that goes through all points in the same subrigion to before it. The value of P_vec does not depend on the looping index so the same quantity was being calculated at each iteration. Now it is calculated only once.

I changed the initialization of some auxiliary points since the logic I was using was not correct. The initialization of p_ovrlp and rotated_p was changed.

I completely changed the was the distance to the torus was calculated. I realized there was a simpler way to do so by just looking at the geometry and reducing it to simpler figures.

I changed the function Rotate_cnt_segment:
1) There is one more variable in the arguments. This new variable was actually the new position of the input point (new_point). Previously I was modifying the input point itself, which led to some inconsistencies. So now the input point is unchanged and the new position is stored in an output variable.
2) To calculate the rotated segment, I needed to move two points (new_point, p_point) to a coordinate system where prev_point was the origin. However, I was not always using those point in the new coordinate system but in the original one. This of course caused the calculations to be incorrect.
3) I modified the way the quadratic equation is solved by working out some more of the algebra.
4) I was taking the minimum positive root to calculate the location of the point rotated_p. However, I actually needed the maximum, so it is the one used now.

Finally, I added some outputs to file for debugging.