Corrected the z-coordinate for ignoring penetrating points
There is a variable for the CNT deposit called zcoord, which is the z-coordinate below which there cannot be penetration of points. That is, any point below this coordinate cannot penetrate a new point.

However, I was not including the van der Waals distance, which resulted in some cases where no penetrating point was found and the point was left hanging. This caused some interpenetration of CNTs.