New section in input file for output data
I added a new section to the input file for output data files. This addition was mainly motivated by the need to have the necessary data to generate nanoparticles in Abaqus. In particular, at the time of this update, GNPs are needed to be generated in Abaqus.

In the particular case in which GNPs are generated in Abaqus as of now, four GNP vertices are needed. Thus there is a flag to generate this file when set to 1. No file is generated when the flag is set to 0. I also implemented the usage of this flag in the code.

I also copied the flag from the Visualization_Flags that generates extra data when mixed or hybrid particles are used. This flags is actually an "output data flag" so it makes sense to have it in that new section. I copied it to make the implement changes in the code in a future update.