Moved flag from Visualization_Flags to Output_Data_Flags
There was a flag in the Visualization_Flags section of the input file that generated extra information when hybrid or mixed particles were used. Since this flag is not actually a visualization flag, I moved it to the section Output_Data_Flags. It is more suitable there. So, I deleted the flag from the Visualization_Flags section of the input file and implemented the changes in the code. Also, deleted the flag from the data structure for visualization flags.