New function to calculate the highest position of a new point
I made a new function to determine the highest position of a new point (when the point is not a seed). This function works iteratively as it tries to find a new direction so that a new point may have a valid position once its highest position is found. A valid position means that the last two CNT segments make an angle at least of PI/2. If no valid position is found, the CNT is rejected. Since now CNTs may be rejected, I added a variable to count the number of rejected CNTs.

I also modified the way the function for the hanging case works. I changed some operations by multiplying by the number resulting by those operations. Now hanging is of PI/6 for the case when there is only one point in the CNT. It used to be PI/4, which seemed a bit steep in some visualizations.

To make the code more consistent with the notation in my notes, I changed the name of a variable from P_vec to v_vec.