Corrected calculation of subregions for CNT deposit
For the CNT deposit I was always calculating a subregion "at floor level". That is, no matter how much the deposit grew up, I was going to always get seeds at the subregions closer to the xy plane.

To change this, now I look for the highest non-empty subregion. In this way, when finding the highest position of the seed, I actually make use of the recursive part of the function that searches in the "layer below".

Because of this change, there is one function for finding the subregion of a seed. For finding the subregion of a CNT point that is not a seed, I used the same function as in the other types of networks. This because for a CNT deposit the cuboids for the sample and extended domain are the same.