Corrected export of GNP vertices in a csv file
The function that exports four vertices of each GNP was modified so that now the coordinates of each vertex is separated by a comma.

I added a new function to convert a Point_3D into a string that does not include parenthesis and takes an integer as input. This integer is the precision, i.e., number of digits after the decimal point. In this way, now we have a function that converts a point into string with a specified precision. Also, this precision is part of the input file and is specified with the flag to export the file with GNP vertices.

I also modified the function that generates this file with the GNP vertices to use the new function that converts a point into a string with a specific precision.