Added generation of file for CNT data output
Added functions to generate a csv file with the coordinates of the CNT points. There is also a file with the number of CNTs, the number of points on each CNT and their radius. This file will help in reading the file with the coordinates to help determine which coordinates depend to which CNT. Also, in case not all CNTs have the same radius this file will help determine the radius of each CNT.