Fiexed invalid anlges between segments
I still had segments that made angles below PI/2. This was because one of the functions that I added to rotate a segment was not being used in the proper case. Even when I modified to be called for the proper case, still did not set the segments in a valid angle. Thus, I decided to not consider the case where the new segment may be rotated to make a valid angle angle. Now the iterative path is always followed.