Added constraints on the location of GNPs
Now GNPs may be generated with the following constraints regarding their location:

1) All GNPs inside the sample
2) At least certain fraction of the GNP volume inside the sample
3) No constraints

To add these options to the input file I reused the first field of the section for GNP_Geometry. This section now will have different meanings depending on the type of network that is being generated. If Hybrid particles are generated, it is used for defining whether CNTs should grow parallel or independent from each other on GNP surfaces. If GNPs or mixed particles are generated, then it is used to determine the constraints on the GNPs locations. This likely will be added to hybrid particles if that code is developed.

Since the input file was modified, I also modified the Input_Reader class to take into account these changes.

I also changed the function that approximates the GNP volume inside the sample. I think the new function is more efficient as it can sometimes break a loop earlier.

In addition, the function that approximates the GNP volume inside the sample was used in a class other than Generate_Network: Cutoff_Wins. I had to modified that class to to accommodate the changes to the function that approximates the GNP volume inside the sample. I also corrected an error where the initial GNP volume was not set correctly.