Modified file for GNP data output
Modified the data in the csv file used to generate GNPs in Abaqus. Now it has the GNP geometry (length in x, length in y, thickness), the rotation angles (in radians) and the location of the centroid.

I also made functions to generate such file and obtain the rotation angles from the rotation matrix.

Temporarily the file for the four vertices approach is still in the code but likely it will be deleted in a future update.