New application for content distribution
I created a new application that calculates the volume fraction at each observation window. With this information, we can observe how the volume fraction changes at each observation window and can be used as an indicator of the distribution of nanoparticles.

Two files are generated: one for CNTs and one for GNPs. To write to file I created a new function in the Printer class that appends a double to a file.

Also, I noticed that there was no check for an invalid Application name in the main function. So I added an else statement to send and error if an invalid application name is given. This was not checked in the Input_Reader class, so it was missing for a long time.

I also commented a line that printed the gnp volume into the output file.