Changed logic when calculating distance to torus
When calculating the distance to the torus, it can be calculated in one of two ways. To decide which way the distance is calculated and if-statement with an AND clause is used. So I changed the AND clause for an OR clause. This change was done because the AND clause requires to always check the two inputs of the clause. The OR statement requires at least one, so the second not always is executed. This results in a potential reduction of computational time.