Correction when recovering angles from rotation matrix
In the rotation matrix the sines of angles have a negative sign. So I used an absolute value to "remove" the sign. However, this was incorrect. What I needed to do is to add a negative sign that cancels out the one in the rotation matrix. This is the correction I made.