Functions to avoid self penetration of CNTs
I added a function to find points within a CNT that the newly generated point penetrates. I also changed the logic for the function that checks for penetration. I removed some check points in the for-loop. I think this makes the logic simpler.

I also found that the function that moved a new point when there are two penetrating points was not working properly. So I made a new version of that function.

For debugging I added new messages to the output file and the possibility to generate visualization files for the generated CNTs in individual files. I also modified the input file to inform of this possibility.