Corrected adding points of new CNT to overlapping subregions
I was adding the points of a new CNT to only one overlapping subregion. However, I was ignoring the fact that these subregions overlap. Thus in some cases I needed to add a point to more than one subregions. Since this was not happening there was still self-penetration. So I added a function to add a point to multiple subregions in the map that I use to detect self-penetration.

I also had to create new functions for avoiding to repeat code. And also noticed that there was some repeated code, which I removed by creating a new function with this repeated code.