New way to add CNT points to global coordinates and found a bug
I added a new function to add CNT points to the global coordinates vector. When adding the CNT points to the global coordinates, all the subregions a point belongs to are calculated. This is also done when filling the map for self penetration. Thus, the new function makes use of the map for self penetration in order to avoid finding the overlapping subregions for each CNT points again.

I also noticed that I had not fill the map for self penetration for the seed point. So I added that and end up finding a bug in the code.

The bug that I found resulted in not grouping correctly the junctions per cluster. All junctions were being grouped into the first cluster. This caused errors for simulations with more than one percolated cluster. I corrected this, and now junctions are grouped per cluster.