Ignoring points at boundaries when exporting data for Abaqus
When creating the CNTs in Abaqus the small segments at the boundary were caused errors when meshing. This because the segment was so small that the sweep was not possible. To avoid this error, I ignore the points at the boundaries. So the function that generates the output data for the CNTs ignores the points at the boundaries.

I noticed that the file with the sample geometry was not being generated. So I added some commands to do this.

I also created a new function that contains all functions used for generating the output data in Generate_Network.