Ignoring CNT points too close to boundaries
When generating the CNT network in Abaqus, it turns out that some CNTs have some portions outside the saple. This causes an error since embedded lements must be inside a host (the sample). This happens because the points are inside the sample but so close that portions of the CNT go outside the sample.

Thus, I added a "percolation layer" which is a CNT distance from the sample's boundary and inside the sample. If a point is in this layer, it is likely that the CNT will make contact with the boundary and result in percolation. Hence the name. So if these points are found, then this is the last CNT point, no need to find a point outside to determine a segment.

I also added a function to the Point_3D class to determine if an instance of the class is inside a cuboid or at its boundary.