Corrected some typos (bugs)
In the Cutoff_Wins class, I was missing a "!" in a function execution within an if-statement. This caused termination due to error when there was actually no error.

In the case of a CNT deposit, when generating the CNTs, there might be points going outside the generation domain (which includes the boundary layer). This causes errors when trying to find the position of the new point as it is not placed in any subregion. Thus, I just added an if-statement to check if this happens and terminate the CNT generation.

Updating the window geometry is done differently in the case of a CNT deposit. So there is a check for this case. However, the check was not being done properly as the string for the CNT deposit was "CNTdeposit" instead of "CNT_deposit". So I just added the "_".