Commented changes regarding percolation layer
The changes related to the percolating layer were causing some numerical errors in the solution: the current that goes into the sample was not the same as the the that goes out.

The changes regarding the percolation layer were added to avoid having CNT elements (and thus, a CNT mesh) outside the host mesh in Abaqus. For now, these changes will be commented as this was causing the numerical errors mentioned above.