Added mass ratio when using  volume fraction and found a bug
When the nanoparticle content is expressed in volume fraction and mixed or hybrid particles are used, now it is possible to specify mass ratio. Before, in such case it was only possible to specify volume ratio. In the case of indicating content as weight fractions, it is still only possible to specify mass ratio.

While testing the change above, I found a bug. After adding the option to restrict the content of GNPs inside the sample, now there are rejected and ignored GNPs. A rejected GNP is a GNP that could not be positioned in a non-penetrating location. An ignored GNP is a GNP that does not have the minimum volume required inside the sample. Now, when checking GNP interpenetration, a GNP might be in a valid psosition, i.e., not rejected but may be ignored. However, every time a GNP was in a valid position, it was added to the subregions it crossed. This resulted in ignored GNPs being added to the subregions. This caused errors. Thus, now a GNP is added to the subregions if it is not ignored (in addition to it not being rejected).