Incorrect update of proper labels for mixed clusters
I found that there was a segmentation error caused when mapping labels after CNT and GNP labels were merged. That is, when assigning proper label number for clusters of mixed particles. The segmentation fault was caused because I was incorrectly doing the clean up of GNP labels (provided there were CNT clusters, otherwise this error does not happen). Here, there was a mismatch between the proper label number and the index in the labels vector because the proper label number was added the number of CNT labels. Thus, proper labels were going outside the range of the labels vector.

So, no I assign labels starting from 0 to GNPs. Then after all (0-based) labels are assigned and clean up I make the adjustment if there are CNT labels.