Corrected the calculation of contact point of GNPs for edge-face case
When two GNPs are in electrical contact, they are separated a certain distance. This distance is also the same distance that separates one simplex on each of the GNPs. In this way, the distance between GNPs may be the distance between to vertices, two edges, a vertex and a face, etc.

So I treat each case separately to identify a single contact point on GNPs that will serve as the point for tunneling. I made a correction for the case edge-face contact. I was not calculating the point correctly. This because there were some cases of the edge-face contact that I did not considered. Actually, now I have a simple solution.

I also added some commented lines for debugging