New penetrating functions for CNTs networks
There is a new function to deal with penetrations so that the length of the CNT segment remains the same. However, the case of only on penetrating point is considered.

There is a version of a function that deals with two penetrating points, but it does not work as the procedure is not entirely correct. In order to implement it, the procedure needs to be revised.

I also changed the logic to place a seed in the non-penetrating case.