Changed content distribution application for CNT deposit
The application to study the content distribution along the observation windows now works differently for a CNT deposit.

First, now the cuboid for the observation window is initialized with the sample cuboid. This is done to initialize the z-coordinate and lengths along the z-direction in the case of a CNT deposit. If this is not done, the quantities along the z-direction will be zero. This change had to be done on the other application (App_Network_3D).

In addition, now I measure the deposited volume per unit area in each observation window instead of the volume fraction in each observation window.

In order to make the change in the previous paragraph, I had to modify some functions that I use to calculate the length of CNT segments inside the sample. The change consisted in using a cuboid instead of the struct Geom_sample. I also added the geom_sample and geom_sample_deposit variables as arguments for the function that generated the CNT deposit.