Dealing with self penetration in the mixed fillers case
Added the functions used to avoid self penetration of CNTs for the case of mixed fillers. Originally I had only implemented these function for the case of only CNTs.

I also found a bug in the function that founds the new position of a point by rotating the CNT segment (One_penetrating_point). I was missing adding the last point in the CNT vector new_cnt to calculate the final position.