Corrected a numerical error
When comparing that a double was less that another double, there was a numerical error when the double were almost equal. This resulted in a true case when the quantities should have been determined to be "equal" (i.e., its difference the cutoff for Zero). To avoid this, I added a Zero (cutoff quantity).