Normal distribution in a range for CNT curvature angle
The angle that determines the CNT curvature was obtained from a normal distribution in a fixed interval. Now that interval can be specified in the input file.

I updated the input file with comments to explain that the range for the normal distribution is required.

I also corrected the function that generated a random number that follows a normal or uniform distribution. Here I changed an if-statement that did not properly compared that two doubles were equal. Also, added an error message in case an invalid distribution type is used.

I also corrected/improved some comments.