Normal distribution in a range for CNT curvature angle in a deposit
I had not changed the generation of the curvature angle for the case of a CNT deposit. I did changes similar to those for randomly distributed CNTs, i.e., I now use the function that generates a random number for a given type of distribution.

Since the variable angle_max is no longer used to generate the curvature angle, I deleted it.

I also added some comments I forgot to add to the previous update.