Progress of generated volume
I realized that there was no progress report when generating GNPs. Thus I added a message every 10% of volume generated, in the same way as it already happens for CNTs. To avoid repeating code and to eliminate some repeated code, I created a function that generates this message. I also improved this function in case more than 20% is increased by adding one particle. This may be more common in the case of GNPs.