Update for calculation of displacement of penetrating GNPs
Originally I calculated a total displacement to move a new GNP that penetrated other GNPs. However, this does not seem to be an accurate way to move the GNP. So I changed the way the final displacement is calculated.

Now, I save the two largest displacements and their corresponding displacement vectors. I apply the displacement vector corresponding to the largest displacement. Then, if there is a second largest displacement, then I apply the perpendicular component to the first displacement.