Patch for when GNP contact is found as edge-triangle
It turns out the GJK does not always return the simplex closest to the origin in the last iteration (in the case of non-penetrating GNPs). Or at least this is the case for my implementation. The GJK is returning a simplex in the Minkowski sum that results in a simplex of 2 vertices and a simplex of 3 vertices in the Cartesian space. The simplex with 3 vertices should not happen and is a direct result of GJK not terminating with the closest simplex to the origin.

For the time being this update deals with such case of a simplex with 3 vertices. These three vertices correspond to only two valid edges. The closest edge to the other simplex (of 2 vertices) is found, provided when translated the edges do intersect.