Corrected edge-edge contact and grouped code into functions
I realized the case for edge-edge contact was not being done properly as it was always assumed edges were colinear (when translated). However, this does not have to be always the case. So now I check if the points are colinear, if so I use the code I had written. It they are not colinear, I wrote new code to handle that case. This new code and the previous code was grouped into a new function to improve readability of the code.

I also created a new function for the edge-face contact to improve readability of the code.