Update for hout command
For some reason, Visual Studio does not recognize commands like "hout<<i" when i is a variable of type size_t. This capability of hout was already implemented but, for some reason I don't remember, it was commented. So I uncommented it and in Visual Studio the error is gone.