Added application to read from Abaqus database to the input file and main function
I added the new application to the main function. So now, the function that will be used to read the Abaqus database is called. However, all the implementation to actually read the database will be implemented in a future update (or updates).

I also added the option to read a network from Abaqus in the input file. In addition, I added to the input file a description of the applications available.