Added usage of variables for odb path and step name
I forgot to use the variables that store the path to the odb file and the step name after being read from the input file. Now they are being used

Also, I added to the input file the format in which a path has to be specified, i.e., with double backslashes.