Removed incorrect usage of window number
For the appliction that read data from an Abaqus database, in two function calls I was using the frame number as the window number. This caused an error when using the window number as a vector index. So Now the window number is set to 0 since only the first window is used when data is read from an Abaqus database.

I also corrected some typos.