Fixed trying to access elements of empty vector
I defined some vectors to store the indices of the points that are dead branches and the backbone. These vectors are then used to export visualization files and only then are the vectors initialized. When no visualization files are exported, these vectors are not initialized. However, I was accessing elements of these vectors even when they were not initialized. This, of course, caused errors. So now, I made sure that elements of these vectors are accessed only when visualization files are exported and, thus, they are initialized.