Ignoring empty CNTs in a shell
When extracting an observation window, an iteration is performed over the CNTs in the shell that corresponds to that observation window.

It may happen that in a previous iteration a CNT is left with less points than the minimum number of points indicated for keeping a CNT. So, that CNT may be ignored and set to contain 0 points. Then in the next iteration, an error ocurrs when extrancting an observation window beacause the points of an empty CNT are bein accessed.

Thus, now, in the iteration for trimming CNTs, I check if a CNT in the shell has points. If it has no points, then it is ignored. Otherwise, the CNT is trimmed.