Removed CNT segments too small for Abaqus
I recalled that the reason I ignored some CNT points when exporting into a csv file was because some points at the boundaries resulted in small segments which Abaqus could not sweep through. The issue with this was that I could not detect percolation when reading from file.

So now, I remove the "small segments" when trimming the CNTs in the boundary layer. For now a "small segment" is a segment that is below a cutoff that for now is set to half the CNT step length. This probably will be better to be added as an input parameter.

In addition, when modifying the functions that trimmed the CNTs in the boundary layer, I found that there were a couple of errors that did not include a "return 0" statement. So I added that statement twice.

Since, the small segments might also appear when extracting the observation window, I made a similar change when CNTs are trimmed when extracting an observation window.

Also, I corrected the number of segments in each CNT, which was not properly calculated. CNTs had one more segment than they should.