Updated flag of CNT points when read from file
I had not updated the flag for the CNT points when they are read from a file. This caused the shells no not be set up properly. So I modified the loop that reads the CNT coordinates so that it is easier to set the flag for each CNT point, where the flag is the CNT number.