Corrected application of displacement to CNTs
Due to the way in which CNTs are generated in Abaqus, the points at the centerline of CNTs are in reverserse order in the node sets comapared to how they are in the CNT structure. Thus I had to apply the displacements in "reverse order".

However, the way I implemented it, I was not applying the displacements in reverse order since I was using the same iterator for the node set and the structure vector. This caused some points to have a final position outside the sample. So I added a new variable to iterate the structure vector in reverse order when iterating over the nodes in the CNT in (the right) order.
This fixed the issue as far as I could test it.
Also added some comments and commands for debugging.