Corrected calculation of new sample geometry and positions of CNTs
I noticed that, to calculate the new sample geometry and new positions of CNTs, I was always adding the displacement at frame i to a given coordinate. However, I did not notice that the displacement at frame i is the total displacement at that frame, no the displacement with respect to the previous frame. In the way I was making the update, I needed the displacement with respect to the previous frame.

Thus, in this update I only add the difference in displacement with respect to the previous frame.