Added a boundary layer along z for CNT deposit
I added a boundary layer to the top boundary along the z-direction in order to increase the generated volume fraction. Given the termination criteria of having a point close to the top boundary (less than a radius). In this way, with the boundary layer on top, the RVE will likely be filled with CNTs.