New class for App that reads network from file
Added the class for a new application that will read the network from a file instead of generating it.

Another class will be added to add the functions thac actually perform the file reading. The class added for this update is only the "application" part, where the resistor network is generated and electrical resistance/resistivity calculated.