New class for reading CNT and GNP data from file
Created a new class that generated a CNT or GNP network by reading data from a file. In the case of CNTs, data may be read from a csv or binary file. In the case of GNPs data may only be red from a csv file, at least for now. In the case of sample geometry it is always read from a csv file since the data is very small and little no nothing is saved in disk space by saving sample geometry data on a binary file.