Added to input file the option to read network from file
The option to read network from file was actually not added to the input file and, thus, was not available to be used. So added the option to read from file and the application that calculates the resistance of the network when read from file.

When the network is read from file, two options are available: comma separated value (csv) and binary (dat). The extension needs to be indicated in the input file. I also modified the function that generates the network from file to use this option instead of the output flag. Because of this, the output flags are not used when reading a network from a file.