Changed loop for multiple samples
The loop that implements the given number of samples is modified so that the object for the resistor network is created and deleted at each iteration (one iteration is one sample). Before, the object was created, all iterations run and lastly the object was deleted. This probably caused memory issues (reading data from a previous sample) that resulted in incorrect results (although not sure this is the cause of the error).