App_Network_3D also reads from file
Added the capability to read a network from file to the application App_Network_3D. This because the only difference between this application and the one I created for reading the file (App_Network_From_File) is the part where the network is gnerated. Thus, application App_Network_From_File is mainly repeated code. In addition this application only works with the option "Read_Network" when defining how the network is generated.

With this update the application App_Network_3D can now generate a random network, a netwrok from seed or from a file (csv or binary).