More options for scope of simulation
What used to be the flag for avoiding calculation of electrical resistance is now the flag for the scope of the simulation. This new name seems more fitting fir its new usage. Now this flag has four integer values from 0 to 3. Each value limits the scope of the simulation:

- 0 means full simulation
- 1 means avoid calculation of electrical resistance
- 2 means avoid calculation of fractions of nanoparticles in the backbone
- 3 means only generate network

These options will give more flexibility according to the what is needed from the code.