Corrected check for the size of the maximum observation window
After reading the dimension of the maximum observation window from the input file, the code checks that those dimensions are valid. Here valid means that the dimensions are positive and that they are not larger than the sample size. However, I was using the incorrect variable when checking that the maximum window size along the z direction is not larger than the sample size. Now it uses the correct variable.