Tamplate for adding displacements to GNPs
Started implementing the function that applies displacements to GNPs. It is a template that considers the cases when a GNP is completely or partially inside the sample. The functions that deal with those cases will be added in a future update.

I also modified the function that finds the GNPs that are partially inside the sample. This function also finds the vertices inside the sample of the GNPs partially inside the sample. This will help in applying displacements to GNPs partially inside the sample.

I also added the function that generates the names for the node sets for the GNPs.