Corrected minor issues
There was a typo in the code when reading the option that a 'csv' files is to be read. The code compared the input against the string 'cvs'. This sent an error to the output file. In addition, this error message was not followed by a 'return 0' statement. This caused the simulation to continue normally. I corrected the typo and added the return statement to terminate the simulation.

When reading from file, there was also the issu of only one observation window being simulated. This was because I originally implemented the functions to read from file to be used exclusively when reading displacements from Abaqus. In such a case, only one observation window is considered. However, now a resistor network simulation may also be run when reading from file. In such a case, multiple observation windows may be considered. Thus, I removed the statement that set the observation window variable to 0 from the class that reads the network from file to the class that applies displacements from Abaqus.