Corrected increase of observation window size when reading from file
When reading from file, there was an issue of the observation windows having large reductions. This was because I originally implemented the functions to read from file to be used exclusively when reading displacements from Abaqus. In such a case, only one observation window is considered and so tehre is no need to increase its size. However, now a resistor network simulation may also be run when reading from file. In such a case, multiple observation windows may be considered. Thus, in the class that reads a network from file, I removed the statements that modified the variables that determine how to update the observation window size.

In addition, I updated some comments and messages in the class Contact_grid. I also added a new variable that is used in several functions.