Corrected vector used to move penetrating GNPs
When moving a new GNP that penetrates another GNP, I was using the unit vector only. This caused a displacemente of 1, which given the units ofthe code is 1 micron. To correct this issue I scale the vector with the penetration depth. When modifying the function that deals with this I forgot to do this step.