Initialized flags for GJK when called in HK76 algorithm
I discovered that I need to initialize the flags for the GJK algorithm even if I don't use them. This is because the GJK implicitly assumes that the flags are initialized as false. Unitialized flags where set to true at execution time. This caused early termination of the GJK and an incorrect calculation of distances between GNPs.

I made other corrections while I got to the realization mentioned above:

1) When reading the input file, it is checked that the lengths of the subregions are smaller than the sample size along each direction.

2) Corrected and added some comments for debugginf in classes Contact_grid and Collision_detection.