GNP-GNP junctions outside the sample are ignored
It turns out there might be GNP-GNP junctions outside the sample. These junctions should be ignored as inside the sample there is no junction. So, I modified the functions that add the GNP-GNP junctions so that when the junction points are outside the sample, the junction is no added to the junction vectors.