GJK version for calculating distance between non-penetrating GNPs
Implemented a version of the GJK that assumes the GNPs are not interpenetrating each other and terminates when finding the separation distance between the GNPs.

The other GJK implementation is more useful as a boolen version. This means that is more useful for determining whether or not the GNPs are interpenetrating each other.

I also added some comments for debuggin to the class for the Hoshen-Kopelman algorithm.