Corrected numerical error with support map and added interpenetration case to GJK_distance
The implementation of the function that obtains the support map for a given direction is the same in Visual Studio and in the code I use for testing the GJK in Eclipse. However, there was one case in which results were different.

The difference in results came from numerical errors when comparing two doubles. Comparing as double1 > double2 resulted in different results than double1 - double2 > Zero, where Zero is a variable to determine a double is small enough to be considered zero. So I changed the implementation to use the second case.

I also added the capability to detect interpenetration in the function for the GJK that calculates distance.