Implemented GJK for distance in Generate_Network class
Implemented the function GJK for distance in the class that generates the GNPs. It is used as a second step after using the boolean GJK.

I also added an error message in case interpenetration is found using GJK_distance. This should not happen as this function is called after the boolean GJK already determined there is no interpenetration.