GJK for distance only
Now the GJK for distance is the only one used since there were some cases when the GJK for distance could not terminate using the simplex output by the boolean GJK.

Also, I did some adjustments to the GJK for distance:
1) The distance from simplex to origin is calculated before starting the while-loop for the GJK.
2) Changed the function that determines if a point is in the simplex. That is, this function used to check if any of the vertices of the simplex was the same as the support point. The new function checks if the support point is conteined in the simplex. That is, if the simplex is an edge (a face) it checks if the support point lies in the line that contains the edge (plane that contains the face).
3) Changed the termination criteria when a point is in the simplex. The condition now uses the function that checks if the point is contained in the simplex. In addition, it is needed that the distance from the simplex to the origin is the same as the length of the projection of the support point (A) onto the direction vector (D).
4) The function that calculates the normal to a tetrahedron face going outside the tetrahedron was changed so that it matches the one I have in Eclipse for testing.
5) In the function of finding the Voroni region for a tetrahedron, the case for the faces was incomplete. Added more conditions that ensure the origin is closest to a face. Incorrect faces were being identified without these additional conditions.
6) Corrected and added comments for debugging.

I also made some changes to the boolean GJK. I changed a "1" to "1.0" when operations with doubles were performed. I also changed a couple of "0" to "Zero".

I changed the precision with which GNP vertices are saved into a vtk file when using the function that exports a single GNP. This is done to reduce numerical errors when testing. I only did it for that function because I use that function to export GNPs that I will use to debug the code. In addition, for files that contain more than one GNP, it will increase the size of those files without being as helpful for debugging as files that contain a single GNP.